Add basic visualizations of polygons using egui_plot (#9)
This change adds the basic infrastructure to visualize polygons using
[`egui_plot`](https://github.com/emilk/egui_plot). This enables running
a local webapp, currently to view the definition of polygons that are
input to tests. These assets are for the moment duplicated in the
visualizer app directory. I'll ultimately want the tests to
automatically output to that directory (or some other one) whatever data
is needed to generate visualizations, which will be a later feature.

- Restructures repo to use workspaces, where source code is in
`computational_geometry` and the visualizer code is in the `visualizer`
directory
- Adds basic webapp code for a visualizer using `egui_plot`