Disable feathering to prevent artifacts in visualizer (#26)
Fixes artifacts in the visualizations of some of the IPA polygons by
disabling feathering in the tesselation which seemed to be the root
cause. Previously you would see long lines shooting out from some of the
triangles in the triangulations, particularly very thin obtuse ones,
those artifacts are gone with this change.