Add extreme point computation from extreme edges (#38)
Adds an implementation of computing extreme edges and extracting extreme
points from extreme edges, as presented in O'Rourke algorithm 3.2.
Differentiates implementations of using this algorithm and computing
extreme points as the complement of interior points. Similarly, adds a
complement function for computing interior points. Also adds aliases
`interior_points` and `extreme_points` functions which resolve to the
current best implementations of each in terms of runtime. Adds a few
convenience helpers and simplifies the test cases.