Refactor ConvexHull to effectively be CCW sorted extreme points (#59)
This change switches `ConvexHull` to have an underlying `HashSet` type
instead of `Vec` to provide protection against duplicates, while
allowing for `Vet` retrieval that retains sorted order of vertices. The
existing hull algorithms and tests are updated to use the types. This
enables just using extreme points in the test cases so that a separate
`convex_hull` attribute doesn't need to be stored, and it has been
removed from all test cases.