Refactor algorithms with computer traits (#60)
This change is a reasonably big refactor to introduce "computer traits",
which effectively make classes of algorithms traits that take `Polygon`
instances as input. This enables drastically reducing the amount of code
on `Polygon` itself, and provides a more ergonomic way to perform
computations in practice and in unit tests.
- Refactors convex hull and triangulation to use computer traits
(functionality pulled out of `Polygon`)
- Removes old benchmarks in favor of new one just for convex hull
- Adds convex hull visualization in place of interior/exterior points
- Moves unit test infrastructure into a `test_util.rs` as common test
infrastructure
- Make convex hull and triangulation tests more compact parametrizing on
algorithms