Add script to modify test case metadata fields across all JSON files (#61)
This change adds a script to batch-process all JSON test case metadata
files to add or remove specified fields. This relieves some tedium when
the `PolygonTestCase` struct changes, it enables easily updating the
corresponding JSON files all at once so that tests pass without having
modify all the files by hand. A simple CLI is provided to specify the
field name and type, and whether to add or remove it. This is a
compromise from a fully automated solution that would sync based on
field introspection of the `PolygonTestCase` struct, which was proving
difficult to do in Rust.