Refactor convex hull to be Polygon (#65)
This change refactors the convex hull algorithms to return a `Polygon`
instance instead of having a separate convex hull type. This is in
preparation to support incremental algorithms that will benefit from the
functionality on the `Polygon` struct, but in trying it out it actually
simplifies things both in the algorithm code and also in the visualizer.