Add implementation of incremental algorithm for computing 2D convex hull (#69)
This change adds an implementation of an Incremental Algorithm for
computing a 2D convex hull for a polygon, as described in O'Rourke
Algorithm 3.7. The approach maintains a growing hull where one new point
is incorporated on each iteration, updating the hull with a similar
approach to Divide and Conquer of finding upper and lower tangents from
the point to the hull and doing a merge step (which excludes points on
the interior). Additionally adds a handful of convenience functions to
`Polygon` and does a full-pass cleanup through the convex hull
algorithms to make things more idiomatic and cleaner.