Add algorithm tracing and visualization (#74)
This change adds basic infrastructure for tracing algorithms and using
the tracer for generating interactive visualizations of the algorithms
in action using the rerun visualizer. When a tracer instance is passed
in to a convex hull algorithm, it will log all intermediate results so
that they can be read out and visualized downstream. Visualization
implementations are added for convex hull using Graham scan, as well as
the incremental algorithm.