Add trace data structures (#83)
This change adds two tracing data structures, a generic `Stack<T>` and
`HullSet`. Both are effectively wrappers of their underlying data
structures (`Vet` and `HashSet`, respectively) but they add
`trace`-level logging when things are added/removed from them. This
provides a significant cleanup in the convex hull algorithms so that all
the trace logging that was in the algorithms themselves is now hidden
away in the calls on the data structure methods. `HullSet` is specific
to tracking `VertexId`s that are considered part of the current convex
hull.